Transferred lots of content from Google Doc to content.md. Used pug mixin for NATO codes. Added binary code to glossary. Made tables and grids
